fix: update imports and typing after restructure
